Copy+pase `viewBuildError.php` logic into controller
